requirements.txt
# Core
python-dotenv>=1.0.0

# AI scoring
openai>=1.0.0
pyyaml

# Supabase
supabase==2.10.0
psycopg2-binary

# HTTP
requests>=2.32.0

# Document text extraction (for CV embeddings)
pypdf>=3.17.0
python-docx>=1.1.0
tiktoken>=0.5.0

# Web server (Render)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0